requirements.txt
# CONNECTION RELATED
requests==2.28.1
flask==2.1.2
gunicorn==20.1.0

# waitress==2.1.2
wheel==0.38.3
Cmake==3.22.4
Werkzeug==2.2.2

# DATABASE 
psycopg2==2.9.3

# TIME
pytz==2022.2.1

# IMAGES
Pillow==9.2.0

# PROFANITY
better_profanity==0.7.0


#APScheduler==3.9.1
#openpyxl==3.0.10

#numpy==1.23.4
#matplotlib==3.6.2
#nsfw-detector==1.1.1
#NudeNet==2.0.9
#tensorflow
#yake==0.4.8
#bcrypt==4.0.1
#bs4==0.0.1